+++ CHANGELOG.md
@@ -16,6 +16,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Create thread now has a body.
 - Made streaming vercel compatible.
 - [frontend] - Make human messages not span the entire width of the chat.
+- Copyright and acknowledgement.
 
 ### Added
 - Docker compose for local development
@@ -37,11 +38,14 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Vitest tests for frontend
 - Title generation endpoint.
 - [frontend] - Dark mode.
+- [frontend] - Automatic title generation
 - [frontend] - Collapsible side bar.
 
 ### Fixed
 - Migrate LLM Evaluation logic to scripts and add tests
+- Query size limiter.
 - Fix /threads/messages endpoint.
+- [frontend] - remove first message delay.
 
 ## [0.3.3] - 30.10.2024